requirements.txt
# LLM 및 AI 모델
openai>=1.0.0
transformers>=4.30.0
torch>=2.0.0
torchvision>=0.15.0

# 한국어 NLP
konlpy>=0.6.0
soynlp>=0.0.493
hgtk>=0.2.1

# 임베딩 및 검색
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
elasticsearch>=8.0.0

# 이미지 처리
Pillow>=9.0.0
opencv-python>=4.8.0

# 데이터 처리
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.3.0

# API 서버
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# 유틸리티
python-dotenv>=1.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
selenium>=4.10.0

# 시각화
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 개발 도구
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0

# 한국어 특화 모델
kobert-transformers>=0.5.1
kcbert>=0.1.0

# LangChain & LangGraph 패키지들
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.0.20

# 벡터 DB & 임베딩 패키지들
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
torch>=1.9.0
transformers>=4.20.0

# 데이터베이스
psycopg2-binary>=2.9.0
sqlite3

# 추가 유틸리티
python-multipart>=0.0.6
typing-extensions>=4.0.0 